requirements.txt
attrs==22.1.0
Automat==20.2.0
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
filelock==3.8.0
Flask==2.2.2
Flask-Compress==1.12
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.4
parsel==1.6.0
plotly==5.10.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
scikit-learn==1.1.2
scipy==1.9.1
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
tldextract==3.3.1
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.12
w3lib==2.0.1
Werkzeug==2.2.2
zipp==3.8.1
zope.interface==5.4.0
